Remove old PROCESS solver implementation
